Add explicit check for age locked mobs
During breeding events, the ages of surrouding mobs were previously changed
without regard to their age lock status, this adds a check for if they
have been age locked to prevent a situation where an age locked baby mob
would become an adult.  Bukkit should probably be honoring age locks
regardless of age, so this is more of a workaround for how ages are
handled than actual fix of something wrong with the original method.